requirements/requirements.txt
coverage
flake8
django-discover-runner
mock
